Added background, moved enemies around, added more coins, added flag
I put a background and moved/added more enemies as well as coins. I also added a flag at the end of the level where if the player collides with it, a win screen pops up. I also added it to where if you fall off the map, you immeditaly die.